Renamed stuff, added regex to cleanup filenames to be used as keys.
