+++ README.md
@@ -1,14 +1,14 @@
 # Stache Parse JSON File
 
-[![npm](https://img.shields.io/npm/v/@blackbaud/skyux-builder-plugin-stache-parse-json-file.svg)](https://www.npmjs.com/package/@blackbaud/skyux-builder-plugin-stache-parse-json-file)
-[![status](https://travis-ci.org/blackbaud/skyux-builder-plugin-stache-parse-json-file.svg?branch=master)](https://travis-ci.org/blackbaud/skyux-builder-plugin-stache-parse-json-file)
+[![npm](https://img.shields.io/npm/v/@blackbaud/skyux-builder-plugin-stache-json-data.svg)](https://www.npmjs.com/package/@blackbaud/skyux-builder-plugin-stache-json-data)
+[![status](https://travis-ci.org/blackbaud/skyux-builder-plugin-stache-json-data.svg?branch=master)](https://travis-ci.org/blackbaud/skyux-builder-plugin-stache-json-data)
 
 This [SKY UX Builder](https://github.com/blackbaud/skyux-builder) plugin takes json files and adds them to a data service to be consumed by components. This plugin is intended to be used with [Stache 2 Components](https://github.com/blackbaud/stache2).
 
 ## Installation
 
 ```
-npm install --save @blackbaud/skyux-builder-plugin-stache-parse-json-file
+npm install --save @blackbaud/skyux-builder-plugin-stache-json-data
 ```
 
 ## Usage
@@ -18,7 +18,7 @@ Open **skyuxconfig.json** and add the following:
 ```
 {
   "plugins": [
-    "@blackbaud/skyux-builder-plugin-stache-parse-json-file"
+    "@blackbaud/skyux-builder-plugin-stache-json-data"
   ]
 }
 ```
